+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add Device Controls (`c`) for running Android devices and iOS Simulators:
+  appearance, text size, and Android locale.
+
 ## [0.8.1] - 2026-07-21
 
 ### Fixed